New evolutionary algorithms class
Evolutionary Algorithms is a huge field per-se, currently maybe shadowed by Deep Learning but likely to come back in vogue as soon as DL limits are reached. Gem `simple_ga` is mainly a demo program, but constitutes a working and usable implementation of a basic Genetic Algorithm. Gem `neuroevo` should (and likely soon will) be divided in two gems as it provides two distinct functionalities: Neural Networks and Evolutionary Strategies (combined in a neuroevolution framework), so it naturally belongs to two classes (my bad on writing a gem with double responsibilities, but it's code I needed at work, and fast).